requirements.txt
pydantic==2.9.2
llama-cpp-python==0.2.90
regex==2024.9.11